use fixed fork of aws-maven project
fixes #3

I deployed https://github.com/ImmobilienScout24/aws-maven to
clojars. This PR contains three commits that fix the
CredentialsProviderChain so that the ProfileCredentialsProvider is used
as the last override to hard-coded credentials.

I've tested the :repository aspect with credentials in .aws/credentials
and it works. I'll be testing that :deploy :library also works shortly.

Hopefully one day the canonical aws-maven repo will get fixed and we can
switch over to that.