Update amazonica
This fixes a problem whereby AWS STS credentials are not usable, e.g. as
issued by aws-vault. The underlying problem seems to in earlier versions
of the aws-java-sdk.

There being no tests to run, I don't know if this breaks anything, but
the limited integrations I've used seem to work fine.